+++ lib/image_input.dart
@@ -1,5 +1,5 @@
 library image_input;
 
 export './widget/image_input.dart';
-export 'package:image_picker/image_picker.dart' show XFile;
+export 'package:image_picker/image_picker.dart' show XFile,ImageSource,CameraDevice;
 export './widget/profile_avatar.dart';