+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.2
+* Added Export of InputSource and CameraDevice from image_picker package
+## 0.0.1
+* Removed setState from library
 ## 0.0.3-beta
 
 * Fixed issue in image_input while selecting image